filter: bail out early if there are no new upstream commits
This avoids resetting upstream/dts to upstream/master and then leaving it there
because there is nothing to do.

+++ scripts/filter.sh
@@ -17,6 +17,12 @@ else
     RANGE="$UPSTREAM_REWRITTEN"
 fi
 
+FETCH_HEAD=$(git rev-parse FETCH_HEAD)
+if [ "$LAST" = "$FETCH_HEAD" ] ; then
+	echo "Nothing new in FETCH_HEAD: $FETCH_HEAD"
+	exit 0
+fi
+
 rm -f .git/refs/original/refs/heads/${UPSTREAM_REWRITTEN}
 
 git branch -f $UPSTREAM_REWRITTEN FETCH_HEAD
